# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Client Onboarding Platform API
  version: 0.4.0-phase-3
  description: Identity, tenant-scoped client/project core, and versioned workflow engine.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate an internal user in one organization
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '202': { description: MFA verification or enrollment is required }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /api/v1/auth/mfa/complete:
    post:
      tags: [Authentication]
      summary: Complete TOTP or one-time recovery-code verification
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeToken, code]
              properties:
                challengeToken: { type: string, maxLength: 200 }
                code: { type: string, maxLength: 40 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Revoke the current session and issue a rotated session token
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auth/forgot-password:
    post:
      tags: [Authentication]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountLookupRequest' }
      responses:
        '202': { description: Generic accepted response }
  /api/v1/auth/reset-password:
    post:
      tags: [Authentication]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenPasswordRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/auth/verify-email:
    post:
      tags: [Authentication]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenPasswordRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auth/accept-invitation:
    post:
      tags: [Authentication]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenPasswordRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auth/email-verification/resend:
    post:
      tags: [Authentication]
      summary: Resend an eligible internal organization invitation without account disclosure
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountLookupRequest' }
      responses:
        '202': { description: Generic accepted response }
...
  /api/v1/projects/{projectId}/onboarding:
    parameters: [{ name: projectId, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      tags: [Onboarding]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
    post:
      tags: [Onboarding]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string, minLength: 8, maxLength: 120 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingStartRequest' } } } }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/onboardings/{onboardingId}:
    get:
      tags: [Onboarding]
      security: [{ cookieAuth: [] }]
      parameters: [{ name: onboardingId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/onboarding-steps/{stepId}/transition:
    post:
      tags: [Onboarding]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - { name: stepId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StepTransitionRequest' } } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: BOS_SESSION
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
  responses:
    Success:
      description: Standard success envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    LoginRequest:
      type: object
      required: [email, password, organizationSlug]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, maxLength: 200 }
        organizationSlug: { type: string, maxLength: 80 }
    AccountLookupRequest:
      type: object
      required: [email, organizationSlug]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        organizationSlug: { type: string, maxLength: 80 }
    TokenPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string, maxLength: 200 }
        password: { type: string, minLength: 12, maxLength: 200 }
...
    OnboardingStartRequest:
      type: object
      additionalProperties: false
      required: [templateVersionId, projectVersion]
      properties:
        templateVersionId: { type: string, format: uuid }
        projectVersion: { type: integer, format: int64, minimum: 0 }
    StepTransitionRequest:
      type: object
      additionalProperties: false
      required: [targetStatus, version]
      properties:
        targetStatus:
          type: string
          enum: [LOCKED, AVAILABLE, IN_PROGRESS, SUBMITTED, UNDER_REVIEW, NEEDS_REVISION, COMPLETED, SKIPPED, FAILED, CANCELLED]
        version: { type: integer, format: int64, minimum: 0 }
    SuccessEnvelope:
      type: object
      required: [success, data, meta, requestId]
      properties:
        success: { const: true }
        data: {}
        meta: { type: object, additionalProperties: true }
        requestId: { type: string, format: uuid }
    ErrorEnvelope:
      type: object
      required: [success, error, requestId]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, fieldErrors]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
        requestId: { type: string, format: uuid }